[int] minor feedback adjustments, mostly readability
